wsj/requirements.txt
html5lib==0.999
aiohttp==1.0.3
tqdm==4.8.4
beautifulsoup4==4.5.1
python_dateutil==2.5.3
